requirements.txt
configparser
confluent_kafka
mysql-connector-python